@@ -1,6 +1,6 @@
 [tool.ruff]
 line-length = 110
-target-version = "py39"
+target-version = "py310"
 
 [tool.ruff.lint]
 fixable = ["ALL"]
@@ -42,7 +42,6 @@ classifiers = [
     'Operating System :: OS Independent',
     'Programming Language :: Python',
     'Programming Language :: Python :: 3',
-    'Programming Language :: Python :: 3.9',
     'Programming Language :: Python :: 3.10',
     'Programming Language :: Python :: 3.11',
     'Programming Language :: Python :: 3.12',
@@ -59,7 +58,7 @@ include = [
 applications-download = "applications_download.cli:main"
 
 [tool.poetry.dependencies]
-python = ">=3.9,<4.0"
+python = ">=3.10,<4.0"
 requests = "2.32.3"
 PyYAML = "6.0.2"
 "ruamel.yaml" = "0.18.10"
