@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
 import argparse
-import sys
+from pathlib import Path
 
 import applications_download
 
@@ -17,13 +17,16 @@ def main() -> None:
             The second file contains the versions of the applications to install,
             this file is usually updated by Renovate."""
     )
-    argparser.add_argument("--applications", help="The file containing the applications to install")
     argparser.add_argument(
-        "--versions", help="The file containing the versions of the applications to install"
+        "--applications", type=Path, help="The file containing the applications to install"
+    )
+    argparser.add_argument(
+        "--versions", type=Path, help="The file containing the versions of the applications to install"
     )
     subparsers = argparser.add_subparsers(dest="cmd")
 
-    subparsers.add_parser("list", help="List the available applications to install")
+    list_parser = subparsers.add_parser("list", help="List the available applications to install")
+    list_parser.add_argument("--name", help="The name of the application to list")
 
     install_parser = subparsers.add_parser("install", help="Install applications")
     install_parser.add_argument("--name", help="The name of the application to install")
@@ -32,7 +35,8 @@ def main() -> None:
     )
     install_parser.add_argument("--all", action="store_true", help="Install all the applications")
 
-    subparsers.add_parser("installed", help="List the installed applications versions")
+    installed_parser = subparsers.add_parser("installed", help="List the installed applications versions")
+    installed_parser.add_argument("--name", help="The name of the application to list")
 
     uninstall_parser = subparsers.add_parser("uninstall", help="Uninstall the installed application")
     uninstall_parser.add_argument("name", help="The name of the application to uninstall")
@@ -41,37 +45,24 @@ def main() -> None:
 
     args = argparser.parse_args()
 
-    applications = applications_download.load_applications(args.applications)
-    versions = applications_download.load_versions(args.versions)
+    applications = applications_download.Applications(args.applications, args.versions)
 
     match args.cmd:
         case "list":
-            for key, app in applications.items():
-                if args.name is None or args.name == key:
-                    version = f" ({versions[key]})" if key in versions else ""
-                    print(f"{key}{version}: {app['description']}")
+            applications.list(args.name)
         case "install":
-            if args.version is not None:
-                if args.name is None:
-                    sys.stderr.write("The --version argument must be used with the --name argument\n")
-                    sys.exit(1)
-                applications_download.install_all_applications(applications, {args.name: args.version})
-            elif args.name is not None:
-                if args.name not in versions:
-                    sys.stderr.write(f"The version of {args.name} is not defined in the versions file\n")
-                    sys.exit(1)
-                applications_download.install_all_applications(applications, {args.name: versions[args.name]})
+            if args.name is not None:
+                applications.install(args.name, args.version)
             elif args.all:
-                applications_download.install_all_applications(applications, versions)
+                applications.install_all()
         case "installed":
-            installed_applications = applications_download.get_installed_applications()
-            for application, version in installed_applications.items():
+            for application, version in applications.installed.items():
                 if args.name is None or args.name == application:
                     print(f"{application}: {version}")
         case "update":
-            applications_download.update_all_applications(applications, versions)
+            applications.update_all()
         case "uninstall":
-            applications_download.uninstall_application(applications, args.name)
+            applications.uninstall(args.name)
         case _:
             argparser.print_help()
 
